+++ NEWS.md
@@ -3,6 +3,7 @@
 - added a function to generate abundance-scaled circles for trends
 - fixed bug preventing tibbles from being passed to grid sampling functions
 - clarified documentation for sampling function
+- fixed bug in `get_species()` for Yellow-bellied Sapsucker
 
 # ebirdst 3.2023.0